Roles

+++ business-logic-service/src/main/java/com/github/evseevda/businesslogicservice/user/mapper/data/RoleJdbcRecordMapper.java
@@ -12,7 +12,7 @@ public class RoleJdbcRecordMapper extends AbstractJdbcRecordMapper<Role> {
 
     @Override
     public Role extractEntity(ResultSet rs) throws SQLException {
-        long id = rs.getLong("id");
+        Integer id = rs.getInt("id");
         String name = rs.getString("name");
         return new Role(id, name);
     }
